requirements.txt
django>=4.0.1,<4.1
gunicorn>=20.1.0,<20.2.0
requests>=2.27.1,<2.28.0
psycopg2-binary>=2.9.3,<3.0.0
APScheduler==3.6.3
asgiref==3.4.1
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.10
dj-database-url==0.5.0
Django==4.0.1
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.3
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
whitenoise==5.3.0
